requirements.txt
django==4.2.7
djangorestframework==3.14.0
psycopg2==2.9.9
drf-spectacular==0.26.5
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
numpy==1.26.2
django-filter==23.5
Pillow==10.2.0